PythonAIAgentFromScratch/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
openai>=1.0.0
python-dotenv>=0.19.0
faiss-cpu>=1.7.4
beautifulsoup4>=4.9.3
requests>=2.26.0
yfinance>=0.1.63